requirements.txt
mysql-connector>=2.2.9
Flask==2.1.0
SQLAlchemy==1.4.18
PyMySQL
bs4>=0.0.1
beautifulsoup4>=4.9.3
selenium>=3.141.0

google-cloud-datastore